+++ CHANGELOG.md
@@ -19,6 +19,7 @@ We go to the next version after each release on [GNOME Shell Extensions website]
 ### Added
 
 - Panel indicator padding size.
+- Window picker close button visibility.
 
 ## [13.0.0 Ring] - 2021-08-10